Initial Commit
My plan is to implement objects using Moo and then allow for multiple
backends (MongoDB, flat file, etc). The first backend will be
implemented in MongoDB. The first object that has been sketched out in
this commit is TimeEntry. This will be the core of this app keeping
track of each change in ones day. The plan is to calculate the
differences between a TimeEntry and the next TimeEntry. This supposes
that the person is working straight in between the two entries. There
will be future support to add/edit entries in between previous entries.