Add last_entry and make collection name a package variable
In the test for `last_entry()`, I needed to add a day to ensure that
the entry just saved would show up as the latest entry. The problem was
caused by the fact that there would be no difference in timestamp
between all the entries created during the test, since it is so fast.
That being the case any of them could show up as the latest entry.