Add timesheet subcommand and move duration calculation to TimeEntry
Since `time_sheet` uses the same duration concept as `current`, I moved
the calculation into `TimeEntry`. If a next entry isn't found, the
duration returned is the current duration.

I think the time_sheet should return an array of hashrefs so that the
timesheet can be rendered however. The timesheet might actually be a
separate object / plugin. Decisions... decisions...