Fix add_entry to use 'local' tz for now
Added test files trying to test for the bug that this fixes but
couldn't figure out how. The scenario was caused by now returning a UTC
DateTime object that hence had a day ahead for it's time. It's day,
month, and year are set in add_entry since the passed time option has
no date component. But when this happens the new datetime with the set
day, etc has it's date and time adjusted again as though it was being
adjusted for the first time.