refact: Move header printing into Timesheet
Previously it was assumed that Timesheet might be used for just the
timesheet portion of a print. For now that doesn't seem likely as the
Timesheet is only used as a complete document and hence needs the
header. In the future this may be refactored again.

This also fixes the lack of date on the first chunk of search results.