feat: Set UTF8 encoding on output
Wrote a test for this by getting all the PerlIO layers and making sure
one of them contains the string `utf8`.

I tried checking that the output contained the correct string, but I
realized this was not testing for the correct issue as the string is
always correct. The issue is the binmode being output not the output
itself.