fix: Only parse ISO8601 dates first if they have a 'T' or 'Z' in them
This issue came up because I would set a manual time (eg `23:00`) via the
command line argument `-t` and sometimes it would set the db entry to be
the wrong date, in some cases it was not just the next day but a little
more than a day. I noticed that this was happening only during the night
ie when it was the next day according to UTC.

Turns out the times were being parsed by the ISO8601 DateTime parser
because it has very lenient rules for what it can parse. The problem was
that it assumed everything was in UTC so set the wrong timezone. So I
put the ISO8601 parser at the end of the list of formats to attempt but
because other formats (in particular `%T`) could parse ISO8601 but
while ignoring the date, it meant that the last test in
`03-time-parsing.t` would fail since `%T` falsely parsed the time
string.

As a compromise, I test to see if a string is explicitly a ISO8601 by
matching against `T` or `Z`. If it matches it parses using ISO8601 first
otherwise it parses with the looser versions first and then uses ISO8601
as a catch all for anything that doesn't parse since it has so many
lenient yet legal options.